Clean up pause system
The player is always processing, unless the pause menu is up, or the death timer has timedout.
The pause menu is always processing unless the game over screen is up.
The pause screen does not come up if the player is playing the dying animation.
Cameras should always follow the SceneTree's pause value.